feat: add use of helper fetch functions to reduce code cluder
